backend/materials/requirements.txt
psycopg2-binary
PyJWT
boto3
openai>=1.0.0
